% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBI700.R
\name{PBI700}
\alias{PBI700}
\title{Production de rapports pour le cours PBI700}
\usage{
PBI700(fichier, dossier = getwd())
}
\arguments{
\item{fichier}{Une suite de caractères définissant le chemin (et le nom) d'un fichier .xlsx qui a été généré automatiquement par un Google Form.}

\item{dossier}{Dossier où les rapports seront générés. Par défaut, le dossier de travail sera utilisé.}
}
\description{
Produit des rapports individuels personnalisés pour chaque étudiant ayant été évalué pour le cours PBI 700 (Séminaires de recherche I de maîtrise) du département de biologie de l'Université de Sherbrooke.
}
\details{
Le fichier doit avoir une structure particulière pour les informations qui sont données aux étudiants soit standard. Spécifiquement, les colonnes avec les noms suivant doivent être incluse dans le fichier :

* Nom de l’étudiant.e qui présente
* Nom de l’évaluateur/l’évaluatrice
* Introduction et mise en contexte - Note
* Introduction et mise en contexte - Commentaires
* Rigueur scientifique - Note
* Rigueur scientifique - Commentaires
* Pédagogie - Note
* Pédagogie - Commentaires
* Réponse aux questions - Note
* Réponse aux questions - commentaires

De plus, toutes les informations doivent être sur la première feuille Excel.
}
