% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBI702.R
\name{PBI702}
\alias{PBI702}
\title{Production de rapports pour les cours PBI 702, PBI 706 et PBI 708}
\usage{
PBI702(fichier, dossier = getwd())
}
\arguments{
\item{fichier}{Une suite de caractères définissant le chemin (et le nom) d'un fichier .xlsx qui a été généré automatiquement par un Google Form.}

\item{dossier}{Dossier où les rapports seront générés. Par défaut, le dossier de travail sera utilisé.}
}
\description{
Produit des rapports individuels personnalisés pour chaque étudiant ayant été évalué pour les séminaires de recherche du département de biologie de l'Université de Sherbrooke.
}
\details{
Le fichier doit avoir une structure particulière pour les informations qui sont données aux étudiants soit standard. Spécifiquement, les colonnes avec les noms suivant doivent être incluse dans le fichier :

* Nom de l’étudiant.e qui présente
* Nom de l’évaluateur/l’évaluatrice
* Rigueur scientifique - Note
* Rigueur scientifique - commentaires
* Pédagogie - Note
* Pédagogie - commentaires
* Réponse aux questions - Note
* Réponse aux questions - commentaires
* Note finale attribuée

De plus, toutes les informations doivent être sur la première feuille Excel.
}
